+++ src/pkg/packager/composer/list.go
@@ -15,6 +15,7 @@ import (
 	"github.com/zarf-dev/zarf/src/extensions/bigbang"
 	"github.com/zarf-dev/zarf/src/pkg/layout"
 	"github.com/zarf-dev/zarf/src/pkg/packager/deprecated"
+	"github.com/zarf-dev/zarf/src/pkg/rules"
 	"github.com/zarf-dev/zarf/src/pkg/utils"
 	"github.com/zarf-dev/zarf/src/pkg/variables"
 	"github.com/zarf-dev/zarf/src/pkg/zoci"
@@ -141,10 +142,9 @@ func NewImportChain(ctx context.Context, head v1alpha1.ZarfComponent, index int,
 			return ic, nil
 		}
 
-		// TODO figure out how to deal with cyclic import before merge
-		// if err := Validate(); err != nil {
-		// 	return ic, err
-		// }
+		if err := rules.ValidateComponent(node.ZarfComponent); err != nil {
+			return ic, err
+		}
 
 		// ensure that remote components are not importing other remote components
 		if node.prev != nil && node.prev.Import.URL != "" && isRemote {
